The DPAV42 `generate_dataset.py` script does not use float16 for saving leading to much larger file size. SO oftimizing file size.

+++ experiments/DPAV42/generate_dataset.py
@@ -172,8 +172,8 @@ def generate_opoi():
     n_profiling = 70000
     n_attack = 10000
 
-    profiling_samples_opoi = np.zeros((len(profiling_samples), 800))
-    attack_samples_opoi = np.zeros((len(attack_samples), 800))
+    profiling_samples_opoi = np.zeros((len(profiling_samples), 800), dtype=np.float16)
+    attack_samples_opoi = np.zeros((len(attack_samples), 800), dtype=np.float16)
 
     for i in range(len(profiling_samples)):
         profiling_samples_opoi[i][:400] = profiling_samples[i][17000:17400]
